chore(deps): bump rusqlite to 0.40
rusqlite 0.40 no longer implements ToSql for u64 (SQLite integers are i64),
so the token-count bindings in add_usage_to_stats cast to i64. Bundled
libsqlite3-sys moves 0.30 -> 0.38.

Verified: pillar-session-store tests (38).

+++ crates/pillar-session-store/src/storage.rs
@@ -996,9 +996,9 @@ pub fn add_usage_to_stats(
              WHERE session_id = ?1",
             rusqlite::params![
                 session_id,
-                cache_read,
-                uncached_tokens,
-                total_tokens,
+                cache_read as i64,
+                uncached_tokens as i64,
+                total_tokens as i64,
                 cost_total
             ],
         )
